Feature/hydrate currency table

+++ server/database/migrations/2021_08_17_102211_create_currencies_table.php
@@ -22,13 +22,13 @@ class CreateCurrenciesTable extends Migration
             $table->decimal("atl_change_percentage")->default(0);
             $table->string("atl_date");
             $table->bigInteger("circulating_supply")->default(0);
-            $table->integer("current_price")->default(0);
-            $table->bigInteger("fully_diluted_valuation")->default(0);
-            $table->bigInteger("high_24h")->default(0);
+            $table->float("current_price", 100, 10)->default(0);
+            $table->float("fully_diluted_valuation", 100, 10)->default(0);
+            $table->float("high_24h", 100, 10)->default(0);
             $table->string("coin_id");
             $table->string("image");
             $table->string("last_updated");
-            $table->integer("low_24h")->default(0);
+            $table->float("low_24h", 100, 10)->default(0);
             $table->bigInteger("market_cap")->default(0);
             $table->float("market_cap_change_24h", 100, 10)->default(0);
             $table->decimal("market_cap_change_percentage_24h")->default(0);
